projrct hexa/requirements.txt
flask==2.0.1
tensorflow==2.8.0
numpy==1.21.0
pillow==8.3.1
opencv-python==4.5.3.56
scikit-learn==0.24.2
gunicorn==20.1.0
python-dotenv==0.19.0 